==> 5_select_instances_to_review.ipynb <==
import pandas as pd

path = "/home/Plaza/mental-health-llm-bias/outputs/"

simple_pronoun = pd.read_csv(path + 'narratives/phq-9_paraphrase_simple_pronoun_replacement.csv')
narrative_pronoun = pd.read_csv(path + 'narratives/phq-9_paraphrase_narrative_pronoun_replacement.csv')
story_pronoun = pd.read_csv(path + 'narratives/phq-9_paraphrase_story_pronoun_replacement.csv')

simple_pronoun_sample = simple_pronoun.groupby('pronoun_replacement').sample(10)[['narrative_raw', 'narrative_paraphrased', 'pronoun_replacement', 'text_w_pronoun']]
simple_pronoun_sample.to_csv(path + 'to_review/phq-9_paraphrase_simple_10_per_pronoun.csv', index=False)

narrative_pronoun_sample = narrative_pronoun.groupby('pronoun_replacement').sample(10)[['narrative_raw', 'narrative_paraphrased', 'pronoun_replacement', 'text_w_pronoun']]
narrative_pronoun_sample.to_csv(path + 'to_review/phq-9_paraphrase_narrative_10_per_pronoun.csv', index=False)

story_pronoun_sample = story_pronoun.groupby('pronoun_replacement').sample(10)[['narrative_raw', 'narrative_paraphrased', 'pronoun_replacement', 'text_w_pronoun']]
story_pronoun_sample.to_csv(path + 'to_review/phq-9_paraphrase_story_10_per_pronoun.csv', index=False)
